+++ gds/src/procedures/similarity/node_similarity.rs
@@ -224,7 +224,7 @@ impl NodeSimilarityFacade {
 
         let pairs: Vec<(u64, u64, f64)> = results
             .iter()
-            .map(|r| (r.source, r.target, r.similarity))
+            .map(|r| (u64::from(r.source), u64::from(r.target), r.similarity))
             .collect();
 
         let updated_store =
